udp mid has to be BIGINT fixes register for longer mid

+++ server/models/udpModel.js
@@ -1,7 +1,7 @@
 import { db, Sequelize } from '../config/db.config'
 
 const Udp = db.define('udp', {
-  mid: Sequelize.INTEGER,
+  mid: Sequelize.BIGINT,
   game: Sequelize.INTEGER,
   usage: Sequelize.INTEGER,
   mode: Sequelize.INTEGER,
